services/api-gateway/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
aiohttp==3.11.11
pyyaml==6.0.2
requests==2.32.3
python-multipart==0.0.9
email-validator==2.1.0
jinja2==3.1.5

# Data Pipeline (Phase 3)
minio==7.2.3
pandas==2.1.4
openpyxl==3.1.2
asyncpg==0.29.0
opensearch-py==2.4.2

# Security & Auth (Phase 5)
PyJWT[crypto]==2.8.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# Semantic Search (Phase 6-7)
sentence-transformers==3.0.0
qdrant-client==1.13.2
numpy>=1.26.0
Pillow>=10.0.0

# ML/AI (Optional for local)
transformers==4.40.2
torch==2.2.0  # Updated for Python 3.12 compatibility

# Task Queue
celery==5.3.6
redis==5.0.1

# Database & Async Utils
greenlet>=3.0.3
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.1
psutil==6.1.1
psycopg2-binary==2.9.9
httpx==0.27.0

# Monitoring
prometheus-client>=0.19.0

# ═══════════════════════════════════════════════════════════════
# НОВI ЗАЛЕЖНОСТI ДЛЯ ТЗ ІНТЕГРАЦІЇ (2025-12-06)
# ═══════════════════════════════════════════════════════════════

# ML/NLP Advanced
spacy==3.7.0               # NER та linguistic features
https://github.com/explosion/spacy-models/releases/download/uk_core_news_sm-3.7.0/uk_core_news_sm-3.7.0-py3-none-any.whl
nlpaug>=1.1.11             # Data augmentation
nltk>=3.8.1                # NLTK for nlpaug
sentencepiece>=0.1.99      # Tokenizer for T5/BART

# Integrations
slack-sdk>=3.27.0          # Slack API client
notion-client>=2.0.0
google-api-python-client>=2.118.0
kubernetes>=29.0.0
prometheus-api-client>=0.5.5
google-auth>=2.25.0        # Google OAuth2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Security Enhancements
argon2-cffi>=23.1.0        # Password hashing (stronger than bcrypt)
cryptography>=41.0.0       # Token encryption

# Observability
structlog>=24.0.0          # WinSURF: Structured Logging
python-json-logger>=2.0.7  # WinSURF: JSON Logs
jaeger-client>=4.8.0       # Distributed tracing (optional)
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-proto==1.30.0
protobuf==5.29.3
importlib-metadata==7.2.1  # Fixed for Aider

# Data Processing
python-docx>=1.1.0         # DOCX parsing for Drive import
pdfplumber>=0.10.0         # PDF parsing
beautifulsoup4>=4.12.3     # HTML parsing
lxml>=5.1.0                # Fast XML/HTML parser

# MLOps & Self-Improvement (v25.0)
dvc[s3]>=3.40.0            # Data Version Control
# augly>=1.0.0               # Facebook AugLy (Optional)
mlflow>=2.10.0             # Experiment Tracking

# LLM Council (v25.0 - Karpathy Pattern)
openai==1.60.2             # OpenAI GPT-4/3.5
anthropic==0.34.0          # Anthropic Claude
google-generativeai==0.8.3 # Google Gemini
groq==0.9.0                # Groq inference
tokenizers==0.19.1         # Fixed for Aider

# Workflow Orchestration
temporalio>=1.4.0

pytest-asyncio>=0.23.0
reportlab>=4.0.0

# Event Bus & Graph
aio-pika==9.4.0
networkx==3.2.1
langgraph==0.0.26

# Telegram v25.0 (aiogram)
aiokafka==0.12.0
aiogram==3.17.0
gTTS==2.5.1
pydub==0.25.1
SpeechRecognition==3.10.1

# ════ WinSURF Golden Core (2025-12-20) ════
duckdb>=0.9.0
pyarrow>=14.0.0
great_expectations>=0.18.0
